requirements.txt
gevent==1.1.1
tornado==4.3
redis
celery[redis]
nose
pymysql
peewee==2.8.1
python-dateutil
arrow
pillow
itsdangerous
httpagentparser
python-geohash
geopy
raven
fabric
qiniu
wtforms==2.0
wtf-peewee
tornado-redis
PyCrypto
elasticsearch
wheezy.captcha
gunicorn
voluptuous
click
aiomysql
xmltodict
Jinja2
jieba
pypinyin
qrcode
pyjwt
fake-factory
cryptography
cached_property
mixer
xlsxwriter
